backend/requirements.txt
argh==0.26.2
asgi-redis==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
autobahn==20.2.1
Automat==20.2.0
autopep8==1.4.4
cffi==1.14.0
channels==2.4.0
Click==7.0
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==3.0.2
django-cors-headers==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack-python==0.5.6
pathtools==0.1.2
pbr==5.4.4
pip-tools==4.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyHamcrest==2.0.0
pylint==2.4.4
pyOpenSSL==19.1.0
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
PyYAML==5.3
redis==2.10.6
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
stevedore==1.31.0
Twisted==19.10.0
txaio==20.1.1
typed-ast==1.4.1
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
watchdog==0.9.0
wrapt==1.11.2
zope.interface==4.7.1
